Relocate pages to pages directory
