+++ frontend/app/App.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import {BrowserRouter, Routes, Route} from 'react-router-dom';
 import './index.css';
-import AddApartment from './Components/AddApartment';
+import AddApartment from './pages/apartment/AddApartment';
 import HomePage from './HomePage';
-import AllApartmnts from './Components/AllApartmnts';
+import AllApartments from './pages/apartment/AllApartmnts';
 import ApartmentView from './Components/ApartmentView';
-import SignUp from './Components/SignUp';
-import LogIn from './Components/LogIn';
-import Messages from './Components/Messages';
-import Profile from './Components/Profile';
+import SignUp from './pages/account/SignUp';
+import LogIn from './pages/account/LogIn';
+import Messages from './pages/messages/Messages';
+import Profile from './pages/account/Profile';
 
 export default function App() {
   return (
@@ -18,7 +18,7 @@ export default function App() {
         <Route path = "*" element={<HomePage/>}/>
         <Route path = "/signup" element={<SignUp/>}/>
         <Route path = "/signin" element={<LogIn/>}/>
-        <Route path = "/allapartments" element={<AllApartmnts/>}/>
+        <Route path = "/allapartments" element={<AllApartments/>}/>
         <Route path = "/apartmntview/:id" element={<ApartmentView/>}/>
         <Route path = "/messages" element={<Messages/>}/>
         <Route path = "/addapartment" element={<AddApartment/>}/>
