+++ frontend/app/pages/account/Profile.js
@@ -1,6 +1,6 @@
-import Navbar from './Navbar';
+import Navbar from '../../components/Navbar';
 import React, {useState, useEffect} from 'react';
-import {auth} from '../FirebaseFuncs';
+import {auth} from '../../FirebaseFuncs';
 
 const Profile = () => {
   const [data, setdata] = useState({
