+++ frontend/app/pages/apartment/AllApartmnts.js
@@ -1,5 +1,5 @@
 import React, {useState, useEffect} from 'react';
-import Navbar from './Navbar';
+import Navbar from '../../components/Navbar';
 import './AllApartmnts.css';
 import testur from '../imgs/the-outpost.jpg';
 import {useNavigate} from 'react-router-dom';
