+++ frontend/app/pages/apartment/ApartmentView.js
